# Develop faster translation from csv to json-ld