… data/schema_org_schemas/HTAN.jsonld
… examples/csv_to_schemaorg.py
… examples/manifest_generator.py