Changed the info returned from explore_class to include only class-specific properties and parents.
- explore_class method in schema/explorer module called methods 
--- retrieving inherited properties from all parent classes of this class
--- retrieving all parents (recursively) of this class
--- this was slow and retrieving too much information given a multidigraph (as opposed to digraph for which it was initially developed). 

Moved a couple of methods from schema/generator to explorer starting a bit of the refactor we discussed with @BrunoGrandePhD (we can discuss during code review).

Made minor changes to manifest generator to avoid unnecessary function calls to google API; @sujaypatil96 we can discuss during code review.

Made minor changes to the csv_utils module to account for DependOn schema header field in @sujaypatil96 'iterative' csv parser version.