@sujaypatil96 2020-12-18
I benchmarked the CSV to JSON-LD conversion process on two schemas (`example.model.csv` and `HIV.model.csv`), and the times taken for the two of them were as follows:

1. `example.model.csv`:
 - old parser: 30.29s
 - new parser: 18.36s

2. `HIV.model.csv`:
 - old parser: 11.07s
 - new parser: 2.72s

Thereby confirming a significant speed-up. I will perform the same analysis on `HTAN.model.csv` and update the results here. But for now, it's good to go.🚀