@milen-sage 2020-12-18
sounds good! yes, the smaller the schema the smaller the speed up; looking forward to your bench on HTAN schema too!