@sujaypatil96 2020-12-18
Comments not critical to this PR, but need to be addressed in the future:

1. Codebase-wide change/transition from using `requiresComponent` to `dependsOnComponent` to ensure consistency/parallel between column/header names in CSV file and corresponding key names in JSON-LD schema (or edge names in networkx graph). For example: change can be made here in `[get_class()](https://github.com/Sage-Bionetworks/schematic/blob/develop/schematic/utils/csv_utils.py#L84)`. Similarly, changes need to be made in other methods like `[explore_class()](https://github.com/Sage-Bionetworks/schematic/blob/develop/schematic/schemas/explorer.py#L169)`, etc.

2. To further improve the speed of the parser the following methods that currently operate/work on the JSON-LD schema can be made to work on the networkx graph instead:
 - `[find_class_specific_properties()](https://github.com/Sage-Bionetworks/schematic/blob/develop/schematic/schemas/explorer.py#L98)`
 - `[find_all_class_properties()](https://github.com/Sage-Bionetworks/schematic/blob/develop/schematic/schemas/explorer.py#L111)`
 - `[find_class_usages()](https://github.com/Sage-Bionetworks/schematic/blob/develop/schematic/schemas/explorer.py#L144)`