@sujaypatil96 2020-12-18
> sounds good! yes, the smaller the schema the smaller the speed up; looking forward to your bench on HTAN schema too!

@milen-sage: Yup, with a few minor changes that I recommended above, I think the PR is ready to be merged.