@milen-sage 2020-12-18
Yes the changes wrt 2) are important for other reasons than speed as well (e.g. having the internal state of schematic completely independent of the input, e.g. json-ld so that we can process potentially other formats (e.g. biolink-ml) relevant for other projects (e.g. INCLUDE).

@sujaypatil96 feel free to merge and open issues for 1 and 2 above; we can prioritize them after the holidays (other refactors and features might take precedence).