@milen-sage 2020-12-18
> `HTAN.model.csv` benchmarking results:
> 
> * old parser: 1609.89s (= 26.83m)
> * new parser: 1164.56s (= 19.4m)
> 
> Approving this PR.🚀

Nice - parser we are replacing > 7 hours